@@ -15,14 +15,15 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@angular/common": "~4.0.0",
-    "@angular/compiler": "~4.0.0",
-    "@angular/core": "~4.0.0",
-    "@angular/forms": "~4.0.0",
+    "@angular/common": "4.4.1",
+    "@angular/compiler": "4.4.1",
+    "@angular/core": "4.4.1",
+    "@angular/forms": "4.4.1",
     "@angular/http": "^4.0.0",
-    "@angular/platform-browser": "~4.0.0",
-    "@angular/platform-browser-dynamic": "~4.0.0",
-    "@angular/router": "~4.0.0",
+    "@angular/platform-browser": "4.4.1",
+    "@angular/platform-browser-dynamic": "4.4.1",
+    "@angular/router": "4.4.1",
+    "@ngrx/effects": "^4.0.5",
     "@ngrx/store": "^4.0.3",
     "@ngrx/store-devtools": "^4.0.0",
     "@ngrx/store-log-monitor": "^3.0.2",
@@ -30,24 +31,24 @@
     "core-js": "^2.4.1",
     "material-design-lite": "^1.3.0",
     "postcss-loader": "^1.3.0",
-    "rxjs": "5.2.0",
+    "rxjs": "5.4.3",
     "zone.js": "^0.8.5"
   },
   "devDependencies": {
     "@types/jasmine": "^2.5.41",
-    "@types/node": "^7.0.5",
+    "@types/node": "8.0.28",
     "@types/protractor": "^4.0.0",
     "@types/selenium-webdriver": "^3.0.0",
     "@types/source-map": "^0.5.0",
     "@types/uglify-js": "^2.6.28",
-    "@types/webpack": "^2.2.5",
+    "@types/webpack": "3.0.10",
     "angular2-template-loader": "^0.6.2",
-    "awesome-typescript-loader": "^3.0.4-rc.0",
+    "awesome-typescript-loader": "3.2.3",
     "bootstrap": "4.0.0-alpha.5",
     "codelyzer": "^2.0.0",
     "css-loader": "^0.27.1",
     "dotenv": "^4.0.0",
-    "extract-text-webpack-plugin": "^1.0.1",
+    "extract-text-webpack-plugin": "1.0.1",
     "file-loader": "^0.10.0",
     "html-loader": "^0.4.4",
     "html-webpack-plugin": "^2.28.0",
