@@ -1,7 +1,7 @@
-import {Http, Headers} from '@angular/http';
-import {Injectable} from '@angular/core';
-import {Store} from '@ngrx/store';
-import {Observable} from "rxjs/Observable";
+import { Http, Headers } from '@angular/http';
+import { Injectable } from '@angular/core';
+import { Store } from '@ngrx/store';
+import { Observable } from "rxjs/Observable";
 import 'rxjs/add/operator/map';
 
 import * as Rx from 'rxjs/Rx'; //
@@ -15,17 +15,15 @@ const BASE_URL = 'http://localhost:5000/phones/';
 export class PhonesService {
   phones: Observable<Array<Phone>>;
 
-  constructor(private http: Http, private store: Store<AppStore>) {
+  constructor(private store: Store<AppStore>) {
     // this.phones = Rx.Observable.combineLatest(
     //   store.select('phones'));
-    this.phones = store.select('phones'); 
+    this.phones = store.select('phones');
   }
 
-  loadItems() {
-    this.http.get(BASE_URL)
-      .map(res => res.json())
-      .map(payload => ({ type: 'ADD_ITEMS', payload }))
-      .subscribe(action => this.store.dispatch(action));
-  }
+  // loadItems() {
+  //   this.http.get(BASE_URL)
+  //     .map(res => res.json());
+  // }
 
 }
