@@ -4,15 +4,9 @@ import { Phone } from './phone.model';
 @Component({
   selector: 'app-list',
   template: `
-  <div *ngFor="let phone of phones" (click)="selected.emit(phone)"
-    class="fem-card mdl-card mdl-shadow--2dp">
-    <div class="mdl-card__title">
-      <h2 class="mdl-card__title-text">{{phone.model}}</h2>
-    </div>
-    <div class="mdl-card__supporting-text">
-      {{phone.branch}}
-    </div>
-  </div>
+  <nav *ngFor="let phone of phones" (click)="selected.emit(phone)" class="mdl-navigation">
+    <a class="mdl-navigation__link">{{phone.model}}</a>
+  </nav>
   `
 })
 export class AppList {
